Replace kernel names
`python3` is the default jupyter kernel name. This jupyterbook must be built
with a kernel with many dependencies (e.g. `allensdk`) installed. In CodeOcean,
we must use a virtual environment to install the deps as they require python
3.8, which is not a commonly available python version (to my knowledge) in
CodeOcean. For whatever reason, it is impossible to shadow the default kernel
name, `python3`, with a custom kernel on CodeOcean. Therefore, I have changed
the kernel name to a name other than the default, specifically `allensdk`.