Fix response latency calculation
We get the time at which the stimulus presentation changes from
`stimulus_presentations`. But if no rows are returned from the table
when we join on change frame, the list of rows are empty, so indexing is
impossible, and we were getting an indexing error.

Now if no rows are returned we just return NaN.