Change CTLUT kernel
The CTLUT code requires the module `hdmf-zarr` which has dependencies
incompatible with the `allensdk`. By using a different ipython kernel
here, we can create a new virtual environment in the CodeOcean capsule
that is separate from the normal kernel with the allensdk and use this
new virtual environment instead to execute the CTLUT code. This keeps
everything separate, avoiding dependency conflicts and other messiness.